+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Rename `TxHash` to `TxID` in `Client.GetTransaction`, `TransactionData`, and `EarnResult`
 - Add Solana options to `Client` methods
 - Mark `SignTransactionRequest.TxHash()` as deprecated in favour of `SignTransactionRequest.TxID()`.
+- Check for duplicate signers for Stellar transactions
 
 ## [v0.2.2](https://github.com/kinecosystem/agora/releases/tag/v0.2.2)
 - Add Kin 2 support